chore: Resolve build warnings (#2632)
* chore: Resolve trivial Kotlin build warnings

* build: Use Gradle test logging events

* chore: Deprecated types

* chore(deps): bump globals from 17.5.0 to 17.6.0 in /frontend

Bumps [globals](https://github.com/sindresorhus/globals) from 17.5.0 to 17.6.0.
- [Release notes](https://github.com/sindresorhus/globals/releases)
- [Commits](https://github.com/sindresorhus/globals/compare/v17.5.0...v17.6.0)

---
updated-dependencies:
- dependency-name: globals
  dependency-version: 17.6.0
  dependency-type: direct:development
  update-type: version-update:semver-minor
...


* chore(deps): bump vue in /frontend in the vue group across 1 directory

Bumps the vue group with 1 update in the /frontend directory: [vue](https://github.com/vuejs/core).

Updates `vue` from 3.5.33 to 3.5.34
- [Release notes](https://github.com/vuejs/core/releases)
- [Changelog](https://github.com/vuejs/core/blob/main/CHANGELOG.md)
- [Commits](https://github.com/vuejs/core/compare/v3.5.33...v3.5.34)

---
updated-dependencies:
- dependency-name: vue
  dependency-version: 3.5.34
  dependency-type: direct:production
  update-type: version-update:semver-patch
  dependency-group: vue
...


* fix: Update GraphQL authentication context handling

* cleanup

---------

+++ app/src/test/kotlin/io/orangebuffalo/simpleaccounting/business/api/expenses/ExpensesQueryTest.kt
@@ -607,7 +607,7 @@ class ExpensesQueryTest(
                                 add(buildJsonObject {
                                     put("node", buildJsonObject {
                                         put("id", testData.expense.id!!.toInt())
-                                        put("version", testData.expense.version!!.toInt())
+                                        put("version", testData.expense.version!!)
                                         put("title", "Slurm delivery invoice")
                                         put("datePaid", "3025-01-15")
                                         put("currency", "EUR")
